+++ controller/konnect/constraints/constraints.go
@@ -26,7 +26,8 @@ type SupportedKonnectEntityType interface {
 		configurationv1alpha1.KongCertificate |
 		configurationv1alpha1.KongTarget |
 		configurationv1alpha1.KongVault |
-		configurationv1alpha1.KongKey
+		configurationv1alpha1.KongKey |
+		configurationv1alpha1.KongKeySet
 	// TODO: add other types
 
 	GetTypeName() string