+++ controller/konnect/watch_kongkeyset.go
@@ -1,0 +1,194 @@
+package konnect
+
+import (
+	"context"
+	"fmt"
+	"reflect"
+
+	"k8s.io/apimachinery/pkg/types"
+	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/builder"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+	"sigs.k8s.io/controller-runtime/pkg/handler"
+	ctrllog "sigs.k8s.io/controller-runtime/pkg/log"
+	"sigs.k8s.io/controller-runtime/pkg/predicate"
+	"sigs.k8s.io/controller-runtime/pkg/reconcile"
+
+	operatorerrors "github.com/kong/gateway-operator/internal/errors"
+	"github.com/kong/gateway-operator/modules/manager/logging"
+
+	configurationv1alpha1 "github.com/kong/kubernetes-configuration/api/configuration/v1alpha1"
+	konnectv1alpha1 "github.com/kong/kubernetes-configuration/api/konnect/v1alpha1"
+)
+
+// KongKeySetReconciliationWatchOptions returns the watch options for the KongKeySet.
+func KongKeySetReconciliationWatchOptions(cl client.Client) []func(*ctrl.Builder) *ctrl.Builder {
+	return []func(*ctrl.Builder) *ctrl.Builder{
+		func(b *ctrl.Builder) *ctrl.Builder {
+			return b.For(&configurationv1alpha1.KongKeySet{},
+				builder.WithPredicates(
+					predicate.NewPredicateFuncs(kongKeySetRefersToKonnectControlPlane),
+				),
+			)
+		},
+		func(b *ctrl.Builder) *ctrl.Builder {
+			return b.Watches(
+				&konnectv1alpha1.KonnectAPIAuthConfiguration{},
+				handler.EnqueueRequestsFromMapFunc(
+					enqueueKongKeySetForKonnectAPIAuthConfiguration(cl),
+				),
+			)
+		},
+		func(b *ctrl.Builder) *ctrl.Builder {
+			return b.Watches(
+				&konnectv1alpha1.KonnectGatewayControlPlane{},
+				handler.EnqueueRequestsFromMapFunc(
+					enqueueKongKeySetForKonnectControlPlane(cl),
+				),
+			)
+		},
+	}
+}
+
+func kongKeySetRefersToKonnectControlPlane(obj client.Object) bool {
+	kongKeySet, ok := obj.(*configurationv1alpha1.KongKeySet)
+	if !ok {
+		ctrllog.FromContext(context.Background()).Error(
+			operatorerrors.ErrUnexpectedObject,
+			"failed to run predicate function",
+			"expected", "KongKeySet", "found", reflect.TypeOf(obj),
+		)
+		return false
+	}
+	return kongKeySet.Spec.ControlPlaneRef != nil &&
+		kongKeySet.Spec.ControlPlaneRef.Type == configurationv1alpha1.ControlPlaneRefKonnectNamespacedRef
+}
+
+func enqueueKongKeySetForKonnectAPIAuthConfiguration(cl client.Client) handler.MapFunc {
+	return func(ctx context.Context, obj client.Object) []reconcile.Request {
+		auth, ok := obj.(*konnectv1alpha1.KonnectAPIAuthConfiguration)
+		if !ok {
+			return nil
+		}
+		var l configurationv1alpha1.KongKeySetList
+		if err := cl.List(ctx, &l, &client.ListOptions{
+			// TODO: change this when cross namespace refs are allowed.
+			Namespace: auth.GetNamespace(),
+		}); err != nil {
+			return nil
+		}
+
+		var ret []reconcile.Request
+		for _, keySet := range l.Items {
+			cpRef, ok := getControlPlaneRef(&keySet).Get()
+			if !ok {
+				continue
+			}
+
+			switch cpRef.Type {
+			case configurationv1alpha1.ControlPlaneRefKonnectNamespacedRef:
+				nn := types.NamespacedName{
+					Name:      cpRef.KonnectNamespacedRef.Name,
+					Namespace: keySet.Namespace,
+				}
+				// TODO: change this when cross namespace refs are allowed.
+				if nn.Namespace != auth.Namespace {
+					continue
+				}
+				var cp konnectv1alpha1.KonnectGatewayControlPlane
+				if err := cl.Get(ctx, nn, &cp); err != nil {
+					ctrllog.FromContext(ctx).Error(
+						err,
+						"failed to get KonnectControlPlane",
+						"KonnectControlPlane", nn,
+					)
+					continue
+				}
+
+				// TODO: change this when cross namespace refs are allowed.
+				if cp.GetKonnectAPIAuthConfigurationRef().Name != auth.Name {
+					continue
+				}
+
+				ret = append(ret, reconcile.Request{
+					NamespacedName: types.NamespacedName{
+						Namespace: keySet.Namespace,
+						Name:      keySet.Name,
+					},
+				})
+
+			case configurationv1alpha1.ControlPlaneRefKonnectID:
+				ctrllog.FromContext(ctx).Error(
+					fmt.Errorf("unimplemented ControlPlaneRef type %q", cpRef.Type),
+					"unimplemented ControlPlaneRef for KongKeySet",
+					"KongKeySet", keySet, "refType", cpRef.Type,
+				)
+				continue
+
+			default:
+				ctrllog.FromContext(ctx).V(logging.DebugLevel.Value()).Info(
+					"unsupported ControlPlaneRef for KongKeySet",
+					"KongKeySet", keySet, "refType", cpRef.Type,
+				)
+				continue
+			}
+		}
+		return ret
+	}
+}
+
+func enqueueKongKeySetForKonnectControlPlane(
+	cl client.Client,
+) func(ctx context.Context, obj client.Object) []reconcile.Request {
+	return func(ctx context.Context, obj client.Object) []reconcile.Request {
+		cp, ok := obj.(*konnectv1alpha1.KonnectGatewayControlPlane)
+		if !ok {
+			return nil
+		}
+		var l configurationv1alpha1.KongKeySetList
+		if err := cl.List(ctx, &l, &client.ListOptions{
+			// TODO: change this when cross namespace refs are allowed.
+			Namespace: cp.GetNamespace(),
+		}); err != nil {
+			return nil
+		}
+
+		var ret []reconcile.Request
+		for _, keySet := range l.Items {
+			cpRef, ok := getControlPlaneRef(&keySet).Get()
+			if !ok {
+				continue
+			}
+			switch cpRef.Type {
+			case configurationv1alpha1.ControlPlaneRefKonnectNamespacedRef:
+				// TODO: change this when cross namespace refs are allowed.
+				if cpRef.KonnectNamespacedRef.Name != cp.Name {
+					continue
+				}
+
+				ret = append(ret, reconcile.Request{
+					NamespacedName: types.NamespacedName{
+						Namespace: keySet.Namespace,
+						Name:      keySet.Name,
+					},
+				})
+
+			case configurationv1alpha1.ControlPlaneRefKonnectID:
+				ctrllog.FromContext(ctx).Error(
+					fmt.Errorf("unimplemented ControlPlaneRef type %q", cpRef.Type),
+					"unimplemented ControlPlaneRef for KongKeySet",
+					"KongKeySet", keySet, "refType", cpRef.Type,
+				)
+				continue
+
+			default:
+				ctrllog.FromContext(ctx).V(logging.DebugLevel.Value()).Info(
+					"unsupported ControlPlaneRef for KongKeySet",
+					"KongKeySet", keySet, "refType", cpRef.Type,
+				)
+				continue
+			}
+		}
+		return ret
+	}
+}